+++ src/main/java/jpabook/jpashop/domain/item/Album.java
@@ -1,6 +1,5 @@
 package jpabook.jpashop.domain.item;
 
-import jpabook.jpashop.domain.Item;
 import lombok.Getter;
 import lombok.Setter;
