+++ src/main/java/jpabook/jpashop/domain/item/Item.java
@@ -1,5 +1,6 @@
-package jpabook.jpashop.domain;
+package jpabook.jpashop.domain.item;
 
+import jpabook.jpashop.domain.Category;
 import lombok.Getter;
 import lombok.Setter;
